requirements.txt
tensorflow-cpu==2.12.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.2
streamlit==1.30.0
Pillow==9.5.0
opencv-python==4.7.0.68
tqdm==4.65.0
